requirements.txt
gradio
huggingface_hub
flask
requests
pillow
werkzeug
gunicorn
